+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Infinite loop when another nvim-ui instance is launched
 - Pass focus to the grid when a new session is launched or connected
 - Occasional crash when closing a session [#46](https://github.com/sakhnik/nvim-ui/issues/46)
+- Data races around global session management
 
 ## [0.0.3] - 2022-05-08